backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.39.0
httpx==0.27.2
langgraph==0.2.34
langchain-core==0.3.10
langchain-anthropic==0.2.3
pydantic==2.9.2
python-dotenv==1.0.1
websockets==13.1
bcrypt==4.2.0
pyjwt==2.9.0
